backend/requirements.txt
langchain
langsmith
langgraph
tiktoken
networkx
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
ipython
langchain-openai
langchain-postgres
langchainhub
pypdf
langchain-community
python-dotenv
